# chore(tests): make tests faster with esbuild, support coverage