--- a/jest.config.ts
+++ b/jest.config.ts
@@ -1,14 +1,8 @@
+import type { Config } from 'jest';
+
 export default {
-  preset: 'ts-jest',
   clearMocks: true,
   displayName: 'WM Extension',
-  collectCoverageFrom: [
-    'src/**/*.test.{js,jsx,ts,tsx}',
-    '!src/**/*.css',
-    '!src/**/*.svg',
-    '!src/**/*.d.ts',
-    '!src/**/index.ts',
-  ],
   coverageDirectory: 'coverage',
   coverageProvider: 'v8',
   maxWorkers: '50%',
@@ -16,6 +10,7 @@ export default {
   moduleNameMapper: {
     '@/(.*)': '<rootDir>/src/$1',
   },
+  passWithNoTests: true,
   setupFilesAfterEnv: ['jest-expect-message', './jest.setup.ts'],
   testMatch: ['**/__tests__/**/*.[jt]s?(x)', '**/?(*.)+(spec|test).[tj]s?(x)'],
   testEnvironment: 'jsdom',
@@ -25,8 +20,7 @@ export default {
     '<rootDir>/jest.config.ts',
   ],
   transform: {
-    '^.+\\.(js|jsx)$': 'babel-jest',
-    '^.+\\.(ts|tsx)?$': 'ts-jest',
+    '^.+\\.(ts|tsx)?$': '@jgoz/jest-esbuild',
     '\\.(css|less|scss|sass|svg)$': 'jest-transform-stub',
   },
-};
+} satisfies Config;