## Context

## Changes proposed in this pull request

- Use `@jgoz/jest-esbuild` to transform ts/tsx files instead of `ts-jest`
- Remove `maxWorkers` restriction from package.json scripts
- Add type checking to jest config
- Fix coverage not picking up files correctly

## Comparison

Running `pnpm test`:

```
Test Suites: 11 passed, 11 total
Tests:       81 passed, 81 total
Snapshots:   0 total
Ran all test suites.
```

### On my PC

```
Before (ts-jest, maxWorkers=2)
Time:        30.298 s

After: with maxWorkers = 2 (and esbuild)
Time:        1.921 s

After: with maxWorkers = 50%  (and esbuild)
Time:        1.867 s
```

### In CI

```
Before:
Time:        21.182 s

After:
Time:        2.674 s
```